+++ lib/src/geom/util/AffineTransformation.dart
@@ -1072,7 +1072,7 @@ class AffineTransformation
   int hashCode() {
     final int prime = 31;
     int result = 1;
-    long temp;
+    int temp;
     temp = Double.doubleToLongBits(m00);
     result = prime * result + (int) (temp ^ (temp >>> 32));
     temp = Double.doubleToLongBits(m01);
